CU80 REFACTOR
- Removed cu80/v2_ansi_rgb & cu80/v2_iso_rgb  sub folders.

*All the below changes occured for both v2_ansi and v2_iso with relevant changes (num leds/ layout etc)*

- Refactor info.json to new format containing most rules.mk and config.h variables
	- Removed variables which now exist in info.json per `qmk info`

- Homoglated keymaps into /default and /via using widows layout (sorry rys :p)

- Updated Readme.mk to includ maz0r as co-maintainer
- ToDo: update build instructions

- Added #ifdef CU80_RGB  in config.h
	- Define's     RGB_MATRIX_KEYPRESSES  &  RGB_MATRIX_MAXIMUM_BRIGHTNESS 50
	- Undefine's      RGBLIGHT_ANIMATIONS, RGBLIGHT_HUE_STEP, RGBLIGHT_SAT_STEP, RGBLIGHT_VAL_STEP, RGBLIGHT_LIMIT_VAL

- Refactor <board>.c added #ifdef block to toggle RGB functionality.

- Added /rgb folder
	- Added config.h
		- toggles CU80_RGB define
		- replaces product ID for VIA
		- replaces product Name for VIA
	- Added rules.mk
		- Sets RGBLIGHT_ENABLE = no       to avoid build errors from having both lighting features enabled.
		- Sets RGB_MATRIX_ENABLE = yes    to build with rgb matrix
		- Defines the RGB_MATRIX_DRIVER = WS2812

>> This results in [WARNINGS] when /rgb firmware is build as the overrides are applied :(

`
QMK Firmware 0.12.15
Making capsunlocked/cu80/v2_ansi with keymap default                                                   [OK]

Making capsunlocked/cu80/v2_ansi with keymap via                                                       [OK]

Making capsunlocked/cu80/v2_ansi/rgb with keymap default                                               [WARNINGS]
	WARNING capsunlocked/cu80/v2_ansi/rgb: PRODUCT_ID in config.h is overwriting usb.pid in info.json
	WARNING capsunlocked/cu80/v2_ansi/rgb: Feature rgblight is specified in both info.json and rules.mk, the rules.mk value wins.

Making capsunlocked/cu80/v2_ansi/rgb with keymap via                                                   [WARNINGS]
	WARNING capsunlocked/cu80/v2_ansi/rgb: PRODUCT_ID in config.h is overwriting usb.pid in info.json
	WARNING capsunlocked/cu80/v2_ansi/rgb: Feature rgblight is specified in both info.json and rules.mk, the rules.mk value wins.

Making capsunlocked/cu80/v2_iso with keymap default                                                    [OK]

Making capsunlocked/cu80/v2_iso with keymap via                                                        [OK]

Making capsunlocked/cu80/v2_iso/rgb with keymap default                                                [WARNINGS]
	WARNING capsunlocked/cu80/v2_iso/rgb: PRODUCT_ID in config.h is overwriting usb.pid in info.json
	WARNING capsunlocked/cu80/v2_iso/rgb: Feature rgblight is specified in both info.json and rules.mk, the rules.mk value wins.

Making capsunlocked/cu80/v2_iso/rgb with keymap via                                                    [WARNINGS]
	WARNING capsunlocked/cu80/v2_iso/rgb: PRODUCT_ID in config.h is overwriting usb.pid in info.json
	WARNING capsunlocked/cu80/v2_iso/rgb: Feature rgblight is specified in both info.json and rules.mk, the rules.mk value wins.
`

- Includes Fixes for Issue #12165 / PR #12166 rgb matrix max brightness.
