+++ keyboards/capsunlocked/cu80/v2_ansi/config.h
@@ -19,37 +19,21 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 #include "config_common.h"
 
-/* USB Device descriptor parameter */
-#define VENDOR_ID       0x4355
-#define PRODUCT_ID      0x0081
-#define DEVICE_VER      0x0001
-#define MANUFACTURER    CapsUnlocked
-#define PRODUCT         CU80 v2 ANSI
-
-/* key matrix size */
-#define MATRIX_ROWS 6
-#define MATRIX_COLS 17
-
-#define MATRIX_ROW_PINS { D3, D2, B7, F1, C7, D5 }
-#define MATRIX_COL_PINS { B0, B1, B2, B3, F4, F5, F6, F7, B6, B5, D7, B4, D6, F0, D1, C6, D4 }
-#define UNUSED_PINS
-
-/* COL2ROW, ROW2COL */
-#define DIODE_DIRECTION COL2ROW
-
-/* Debounce reduces chatter (unintended double-presses) - set 0 if debouncing is not needed */
-#define DEBOUNCE 5
-
 #define LOCKING_SUPPORT_ENABLE
 /* Locking resynchronize hack */
 #define LOCKING_RESYNC_ENABLE
 
-#define RGB_DI_PIN E6
-#ifdef RGB_DI_PIN
-    #define RGBLIGHT_ANIMATIONS
-    #define RGBLED_NUM 87
-    #define RGBLIGHT_HUE_STEP 8
-    #define RGBLIGHT_SAT_STEP 8
-    #define RGBLIGHT_VAL_STEP 8
-    #define RGBLIGHT_LIMIT_VAL 50
+/* Define RGB */
+#define DRIVER_LED_TOTAL 87
+
+/* RGB firmware override */
+#ifdef CU80_RGB
+    #define RGB_MATRIX_KEYPRESSES
+    #define RGB_MATRIX_MAXIMUM_BRIGHTNESS 50
+    #undef RGBLIGHT_ANIMATIONS
+    #undef RGBLIGHT_HUE_STEP
+    #undef RGBLIGHT_SAT_STEP
+    #undef RGBLIGHT_VAL_STEP
+    #undef RGBLIGHT_LIMIT_VAL
 #endif
+
