+++ keyboards/capsunlocked/cu80/v2_ansi/keymaps/via/rules.mk
@@ -1,1 +1,2 @@
 VIA_ENABLE = yes
+VIAL_ENABLE = yes
