+++ keyboards/capsunlocked/cu80/v2_ansi/rgb/config.h
@@ -1,0 +1,9 @@
+/* enable RGB matrix */
+#define CU80_RGB
+
+/*update the product for VIA/VIAL and Remap-keys.app */
+#undef PRODUCT_ID
+#undef PRODUCT
+
+#define PRODUCT_ID      0x0082
+#define PRODUCT         CU80 v2 ANSI RGB
