--- keyboards/capsunlocked/cu80/v2_ansi_rgb/config.h
@@ -1,55 +1,0 @@
-/*
-Copyright 2021 CapsUnlocked
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 2 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-#pragma once
-
-#include "config_common.h"
-
-/* USB Device descriptor parameter */
-#define VENDOR_ID       0x4355
-#define PRODUCT_ID      0x0082
-#define DEVICE_VER      0x0001
-#define MANUFACTURER    CapsUnlocked
-#define PRODUCT         CU80 v2 ANSI RGB
-
-/* key matrix size */
-#define MATRIX_ROWS 6
-#define MATRIX_COLS 17
-
-#define MATRIX_ROW_PINS { D3, D2, B7, F1, C7, D5 }
-#define MATRIX_COL_PINS { B0, B1, B2, B3, F4, F5, F6, F7, B6, B5, D7, B4, D6, F0, D1, C6, D4 }
-#define UNUSED_PINS
-
-/* COL2ROW, ROW2COL */
-#define DIODE_DIRECTION COL2ROW
-
-/* Debounce reduces chatter (unintended double-presses) - set 0 if debouncing is not needed */
-#define DEBOUNCE 5
-
-#define LOCKING_SUPPORT_ENABLE
-/* Locking resynchronize hack */
-#define LOCKING_RESYNC_ENABLE
-
-
-/* Define RGB */
-#define RGB_DI_PIN E6
-#define RGB_MATRIX_KEYPRESSES
-#define DRIVER_LED_TOTAL 87
-#define RGBLED_NUM 87
-#define RGBLIGHT_LIMIT_VAL 50
-
-
