--- keyboards/capsunlocked/cu80/v2_ansi_rgb/info.json
@@ -1,100 +1,0 @@
-{
-    "keyboard_name": "CapsUnlocked CU80 v2 ANSI RGB",
-    "maintainer": "rys",
-    "url": "https://caps-unlocked.com/cu80-round-2/",
-    "width": 18.25,
-    "height": 6.5,
-    "layouts": {
-        "LAYOUT_tkl_ansi": {
-            "layout": [
-                {"label":"Esc", "x":0, "y":0},
-                {"label":"F1", "x":2, "y":0},
-                {"label":"F2", "x":3, "y":0},
-                {"label":"F3", "x":4, "y":0},
-                {"label":"F4", "x":5, "y":0},
-                {"label":"F5", "x":6.5, "y":0},
-                {"label":"F6", "x":7.5, "y":0},
-                {"label":"F7", "x":8.5, "y":0},
-                {"label":"F8", "x":9.5, "y":0},
-                {"label":"F9", "x":11, "y":0},
-                {"label":"F10", "x":12, "y":0},
-                {"label":"F11", "x":13, "y":0},
-                {"label":"F12", "x":14, "y":0},
-                {"label":"PrtSc", "x":15.25, "y":0},
-                {"label":"Scroll Lock", "x":16.25, "y":0},
-                {"label":"Pause", "x":17.25, "y":0},
-                {"label":"\u00ac", "x":0, "y":1.5},
-                {"label":"!", "x":1, "y":1.5},
-                {"label":"@", "x":2, "y":1.5},
-                {"label":"£", "x":3, "y":1.5},
-                {"label":"$", "x":4, "y":1.5},
-                {"label":"%", "x":5, "y":1.5},
-                {"label":"^", "x":6, "y":1.5},
-                {"label":"&", "x":7, "y":1.5},
-                {"label":"*", "x":8, "y":1.5},
-                {"label":"(", "x":9, "y":1.5},
-                {"label":")", "x":10, "y":1.5},
-                {"label":"_", "x":11, "y":1.5},
-                {"label":"+", "x":12, "y":1.5},
-                {"label":"Backspace", "x":13, "y":1.5, "w":2},
-                {"label":"Insert", "x":15.25, "y":1.5},
-                {"label":"Home", "x":16.25, "y":1.5},
-                {"label":"PgUp", "x":17.25, "y":1.5},
-                {"label":"Tab", "x":0, "y":2.5, "w":1.5},
-                {"label":"Q", "x":1.5, "y":2.5},
-                {"label":"W", "x":2.5, "y":2.5},
-                {"label":"E", "x":3.5, "y":2.5},
-                {"label":"R", "x":4.5, "y":2.5},
-                {"label":"T", "x":5.5, "y":2.5},
-                {"label":"Y", "x":6.5, "y":2.5},
-                {"label":"U", "x":7.5, "y":2.5},
-                {"label":"I", "x":8.5, "y":2.5},
-                {"label":"O", "x":9.5, "y":2.5},
-                {"label":"P", "x":10.5, "y":2.5},
-                {"label":"{", "x":11.5, "y":2.5},
-                {"label":"}", "x":12.5, "y":2.5},
-                {"label":"|", "x":13.5, "y":2.5, "w":1.5},
-                {"label":"Delete", "x":15.25, "y":2.5},
-                {"label":"End", "x":16.25, "y":2.5},
-                {"label":"PgDn", "x":17.25, "y":2.5},
-                {"label":"Caps Lock", "x":0, "y":3.5, "w":1.75},
-                {"label":"A", "x":1.75, "y":3.5},
-                {"label":"S", "x":2.75, "y":3.5},
-                {"label":"D", "x":3.75, "y":3.5},
-                {"label":"F", "x":4.75, "y":3.5},
-                {"label":"G", "x":5.75, "y":3.5},
-                {"label":"H", "x":6.75, "y":3.5},
-                {"label":"J", "x":7.75, "y":3.5},
-                {"label":"K", "x":8.75, "y":3.5},
-                {"label":"L", "x":9.75, "y":3.5},
-                {"label":":", "x":10.75, "y":3.5},
-                {"label":"\"", "x":11.75, "y":3.5},
-                {"label":"Enter", "x":12.75, "y":3.5, "w":2.25},
-                {"label":"Shift", "x":0, "y":4.5, "w":2.25},
-                {"label":"Z", "x":2.25, "y":4.5},
-                {"label":"X", "x":3.25, "y":4.5},
-                {"label":"C", "x":4.25, "y":4.5},
-                {"label":"V", "x":5.25, "y":4.5},
-                {"label":"B", "x":6.25, "y":4.5},
-                {"label":"N", "x":7.25, "y":4.5},
-                {"label":"M", "x":8.25, "y":4.5},
-                {"label":"<", "x":9.25, "y":4.5},
-                {"label":">", "x":10.25, "y":4.5},
-                {"label":"?", "x":11.25, "y":4.5},
-                {"label":"Shift", "x":12.25, "y":4.5, "w":2.75},
-                {"label":"Up", "x":16.25, "y":4.5},
-                {"label":"Ctrl", "x":0, "y":5.5, "w":1.25},
-                {"label":"Alt", "x":1.25, "y":5.5, "w":1.25},
-                {"label":"Win", "x":2.5, "y":5.5, "w":1.25},
-                {"label":"Space", "x":3.75, "y":5.5, "w":6.25},
-                {"label":"Win", "x":10, "y":5.5, "w":1.25},
-                {"label":"AltGr", "x":11.25, "y":5.5, "w":1.25},
-                {"label":"Ctrl", "x":12.5, "y":5.5, "w":1.25},
-                {"label":"Fn", "x":13.75, "y":5.5, "w":1.25},
-                {"label":"Left", "x":15.25, "y":5.5},
-                {"label":"Down", "x":16.25, "y":5.5},
-                {"label":"Right", "x":17.25, "y":5.5}
-            ]
-        }
-    }
-}
