--- keyboards/capsunlocked/cu80/v2_ansi_rgb/keymaps/default/readme.md
@@ -1,3 +1,0 @@
-# The default ANSI keymap for the CU80
-
-This CU80 is keymap is for the fixed ANSI variant with fully populated bottom row.
