+++ keyboards/capsunlocked/cu80/v2_iso/rgb/config.h
@@ -1,0 +1,8 @@
+/* enable RGB matrix */
+#define CU80_RGB
+
+/*update the product for VIA/VIAL and Remap-keys.app */
+#undef PRODUCT_ID
+#define PRODUCT_ID      0x0083
+#undef PRODUCT
+#define PRODUCT         CU80 v2 ISO RGB
