+++ keyboards/capsunlocked/cu80/v2_iso/rgb/rules.mk
@@ -1,0 +1,3 @@
+RGBLIGHT_ENABLE = no       # Enable keyboard RGB underglow
+RGB_MATRIX_ENABLE = yes     # Enable RGB matrix
+RGB_MATRIX_DRIVER = WS2812  # Set Driver to WS2812
