--- keyboards/capsunlocked/cu80/v2_iso_rgb/keymaps/default/readme.md
@@ -1,3 +1,0 @@
-# The default ISO keymap for the CU80
-
-This CU80 is keymap is for the fixed ISO variant with fully populated bottom row.
