--- keyboards/capsunlocked/cu80/v2_iso_rgb/keymaps/via/rules.mk
@@ -1,1 +1,0 @@
-VIA_ENABLE = yes
